requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
sqlalchemy==2.0.48
asyncpg==0.31.0
python-jose==3.5.0
passlib==1.7.4
pydantic-settings==2.13.1
python-dotenv==1.2.2
cryptography==46.0.3
python-multipart==0.0.22
websockets==16.0
